requirements.txt
asgiref==3.7.2
Django==4.2.5
django-advance-thumbnail==1.0.0
django-appconf==1.0.5
django-imagekit==5.0.0
djangorestframework==3.14.0
pilkit==3.0
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
